Move internal files to internal directory
