+++ internal/download.go
@@ -1,4 +1,4 @@
-package disclosure
+package internal
 
 import (
 	"io"
